requirements.txt
python-telegram-bot==20.3
openai==1.12.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
loguru==0.7.2